refactor: move manager to separate module
- Move manager logic to own module
- Move async logic to own module
- Improve type annotations
